Allow parsing of large files

+++ OriginAnyParser.cpp
@@ -898,7 +898,7 @@ void OriginAnyParser::readAttachmentList() {
 	att_header = string(12,0);
 	while (true) {
 		// check for eof
-		if (((unsigned int)file.tellg() == d_file_size) || (file.eof())) break;
+		if ((file.tellg() == d_file_size) || (file.eof())) break;
 		// cannot use readObjectAsString: there is no '\n' at end
 		file.read(reinterpret_cast<char*>(&att_header[0]), 12);
